+++ ExaminationSystem/Services/CourseService.cs
@@ -2,7 +2,9 @@
 using AutoMapper.QueryableExtensions;
 using ExaminationSystem.DTOs.CourseDTOs;
 using ExaminationSystem.Models;
+using ExaminationSystem.Models.Enums;
 using ExaminationSystem.Repositories.Implementations;
+using Microsoft.EntityFrameworkCore;
 
 namespace ExaminationSystem.Services;
 
@@ -10,27 +12,30 @@ public class CourseService
 {
     BaseRepository<Course> _courseRepo;
     BaseRepository<Instructor> _instructorRepo;
-    BaseRepository<StudentCourse> _studentCourseRepo;
-    BaseRepository<Student> _studentRepo;
+    BaseRepository<Exam> _examRepo;
     IMapper _mapper;
     public CourseService(IMapper mapper)
     {
         _courseRepo = new BaseRepository<Course>();
         _instructorRepo = new BaseRepository<Instructor>();
-        _studentCourseRepo = new BaseRepository<StudentCourse>();
-        _studentRepo = new BaseRepository<Student>();
+        _examRepo = new BaseRepository<Exam>();
         _mapper = mapper;
     }
 
-    public IEnumerable<CourseDetailsDTO> GetAllCourses()
+    public async Task<IEnumerable<CourseDetailsDTO>> GetAllCoursesAsync(int istructorId)
     {
-        var courses = _courseRepo.GetAll()
+        var courses = await _courseRepo.GetAll().Where(m => m.CreatedBy == istructorId)
                                  .ProjectTo<CourseDetailsDTO>(_mapper.ConfigurationProvider)
                                  .OrderBy(m => m.Name)
-                                 .ToList();
+                                 .ToListAsync();
+
+        if (!courses.Any())
+            throw new Exception("Not found any courses");
+
         return courses;
     }
-    public async Task<CourseDetailsDTO> AddCourseAsync(CreateCourseDTO dto)
+
+    public async Task<CourseDetailsDTO> AddCourseAsync(CreateCourseDTO dto, int instructorID)
     {
         var instructor = await _instructorRepo.GetByIdAsync(dto.InstructorID);
 
@@ -38,6 +43,7 @@ public class CourseService
             throw new Exception($"No instructor was found with ID = {dto.InstructorID}");
 
         var course = _mapper.Map<Course>(dto);
+        course.CreatedBy = instructorID;
 
         await _courseRepo.AddAsync(course);
 
@@ -48,24 +54,30 @@ public class CourseService
         return courseDetails;
     }
 
-    public async Task<CourseDetailsDTO> GetCourseByIDAsync(int id)
+    public async Task<CourseDetailsDTO> GetCourseByIDAsync(int courseID, int instructorID)
     {
-        var course = await _courseRepo.GetByIdAsync(id, c => c.Instructor);
+        var course = await _courseRepo.GetByIdAsync(courseID, c => c.Instructor);
 
         if (course == null)
-            throw new Exception($"No course was found with ID = {id}");
+            throw new Exception($"No course was found with ID = {courseID}");
+
+        if (course.InstructorID != instructorID)
+            throw new UnauthorizedAccessException("You can not view this course");
 
         var courseDetails = _mapper.Map<CourseDetailsDTO>(course);
 
         return courseDetails;
     }
 
-    public async Task<CourseDetailsDTO> UpdateCourseAsync(int courseID, UpdateCourseDTO updateCourseDTO)
+    public async Task<CourseDetailsDTO> UpdateCourseAsync(int courseID, int instructorID, UpdateCourseDTO updateCourseDTO)
     {
-        var courseIsExist = await _courseRepo.AnyAsync(m => m.ID == courseID);
+        var course = await _courseRepo.GetByIdAsync(courseID);
 
-        if (!courseIsExist)
-            throw new Exception($"Not Found Course With ID {courseID}");
+        if (course == null)
+            throw new Exception($"Not found course with ID {courseID}");
+
+        if (course.InstructorID != instructorID)
+            throw new UnauthorizedAccessException("You can not update this course.");
 
         var result = await _courseRepo.UpdateAsync(
         c => c.ID == courseID,
@@ -73,51 +85,48 @@ public class CourseService
               .SetProperty(d => d.Name, updateCourseDTO.Name)
               .SetProperty(d => d.Description, updateCourseDTO.Description)
               .SetProperty(d => d.Hours, updateCourseDTO.Hours)
-              .SetProperty(d => d.InstructorID, updateCourseDTO.InstructorID)
               .SetProperty(d => d.UpdatedAt, DateTime.UtcNow)
+              .SetProperty(d => d.UpdatedBy, instructorID)
         );
 
         if (result == 0)
             throw new Exception("Update failed");
 
-        return await GetCourseByIDAsync(courseID);
+        var newCourse = await _courseRepo.GetByIdAsync(courseID, c => c.Instructor);
+
+        var courseDetailsDto = _mapper.Map<CourseDetailsDTO>(newCourse);
+
+        return courseDetailsDto;
     }
 
-    public async Task DeleteCourse(int id)
+    public async Task DeleteCourse(int id,int instructorID)
     {
         var course = await _courseRepo.GetByIdAsync(id);
 
         if (course == null)
             throw new Exception($"No course was found with ID = {id}");
 
+        if(course.CreatedBy != instructorID)
+            throw new UnauthorizedAccessException("You can not delete this course");
+
         await _courseRepo.DeleteAsync(course);
     }
 
-    public async Task DeleteCourseSoftAsync(int id)
+    public async Task DeleteCourseSoftAsync(int courseID, int instructorID)
     {
-        var result = await _courseRepo.SoftDeleteAsync(id);
+        var course = await _courseRepo.GetByIdAsync(courseID);
+
+        if (course == null)
+            throw new Exception($"Not found course with ID {courseID}");
+
+        if (course.CreatedBy != instructorID)
+            throw new UnauthorizedAccessException("You can not delete this course");
+
+        var result = await _courseRepo.SoftDeleteAsync(courseID, instructorID);
 
         if (result == 0)
             throw new Exception("Deleted was failed");
     }
 
-    public async Task AssignStudentToCourse(EnrollInCourseDTO dto)
-    {
-        var isCourseExist = await _courseRepo.AnyAsync(c => c.ID == dto.CourseID);
 
-        if (!isCourseExist)
-            throw new Exception("This Course is Not Exist.");
-
-        var isStudentExist = await _studentRepo.AnyAsync(s => s.ID == dto.StudentID);
-
-        if (!isStudentExist)
-            throw new Exception("This Student is Not Exist.");
-
-        var isStudentEnrolled = await _studentCourseRepo.AnyAsync(m => m.StudetnID == dto.StudentID && m.CourseID == dto.CourseID);
-
-        if (isStudentEnrolled)
-            throw new Exception("This Student is Already Enrolled.");
-
-        await _studentCourseRepo.AddAsync(new StudentCourse { StudetnID = dto.StudentID, CourseID = dto.CourseID });
-    }
 }
